Move shared controller methods into concern
